refactor: 将 AgentSettings 迁移至独立的 agents 模块
- 将 AgentSettings 从 pages/settings 迁至 pages/agents，使模块职责更清晰
- 更新 GuidPage 和路由中对应 import 路径
- 清理 skills 模块中冗余的 license header 注释
