+++ src/renderer/pages/skills/components/SkillAuditReportModal.tsx
@@ -1,9 +1,3 @@
-/**
- * @license
- * Copyright 2025 Sudowork (sudowork.ai)
- * SPDX-License-Identifier: Apache-2.0
- */
-
 import React from 'react';
 import { Modal } from '@arco-design/web-react';
 import { X, Shield } from 'lucide-react';
