+++ src/renderer/pages/skills/components/SkillAuditSummary.tsx
@@ -1,9 +1,3 @@
-/**
- * @license
- * Copyright 2025 Sudowork (sudowork.ai)
- * SPDX-License-Identifier: Apache-2.0
- */
-
 import React, { useEffect, useState } from 'react';
 import { Button, Spin } from '@arco-design/web-react';
 import { IconFolder } from '@arco-design/web-react/icon';
