+++ src/renderer/pages/skills/components/SkillCard.tsx
@@ -1,9 +1,3 @@
-/**
- * @license
- * Copyright 2025 Sudowork (sudowork.ai)
- * SPDX-License-Identifier: Apache-2.0
- */
-
 import React from 'react';
 import { Button, Progress, Tooltip } from '@arco-design/web-react';
 import { Download } from 'lucide-react';
