+++ src/renderer/pages/skills/index.tsx
@@ -1,9 +1,3 @@
-/**
- * @license
- * Copyright 2025 Sudowork (sudowork.ai)
- * SPDX-License-Identifier: Apache-2.0
- */
-
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { Button, Spin, Message, Input, Modal, Tooltip } from '@arco-design/web-react';
 import { IconSearch } from '@arco-design/web-react/icon';
