+++ src/renderer/pages/skills/types/index.ts
@@ -1,9 +1,3 @@
-/**
- * @license
- * Copyright 2025 Sudowork (sudowork.ai)
- * SPDX-License-Identifier: Apache-2.0
- */
-
 import type { ISkillHubDetail } from '@/common/ipcBridge';
 
 export interface IBridgeResponse<D = unknown> {
